Refactor mutations: remove hooks
We will rarely be using the state provided from the useMutation hook, so using a hook here would be unnecessary. This solution makes mutations simpler and also makes it easier to replace them with another technology should we want to.
In a rare case that we do want state to be maintained, we can manually do this fairly trivially.